исправлены ошибки в функции для парсинга картинок добавлена тоблица для записи и чтения списка ip,

+++ bot.py
@@ -2,8 +2,6 @@ from random import randint
 import asyncio
 import os
 
-import time
-
 import psycopg2
 
 from googletrans import Translator
@@ -28,16 +26,19 @@ dp = Dispatcher(bot)
 
 import pandas as pd
 
-cols = [1, 2, 3, 4, 5, 6, 7]
+cols_user_base = [1, 2, 3, 4, 5, 6, 7]
+cols_ip_base = [1, 2]
 global ch_log, con_adm_rig, ip_list, element
-element = 0
-ip_list = []
 ch_log = 0
 con_adm_rig = 0
 Im_sear = 0
-user_base = pd.read_excel('user_base.xlsx', usecols=cols)
+user_base = pd.read_excel("User_base.xlsx", usecols=cols_user_base)
 user_base.head()
-
+ip_base = pd.read_excel('ip_base.xlsx', usecols=cols_ip_base)
+user_base.head()
+ip = ip_base['number_ip'].tolist()
+element = int(ip[0])
+ip_list = ip_base['ip_list'].tolist()
 
 @dp.message_handler(commands="dice")
 async def cmd_random(message: types.Message):
@@ -57,18 +58,19 @@ async def send_random_value(call: types.CallbackQuery):
 @dp.message_handler(commands=['start'])
 async def process_start_command(message: types.Message):
     global user_base
+    user_id_base = user_base['id'].tolist()
     user_id = message.from_user.id
-    if ban_chek(user_id) != 1:
-        user_id_base = user_base['id'].tolist()
+    if ban_chek(user_id) != 1 or (user_id in user_id_base) == False:
         user_name_base = user_base['name'].tolist()
         if (user_id in user_id_base) == False:
-            await bot.send_message(message.from_user.id,'Здавствуй пользователь я бот Grek, как я вижу ты здесь впервые укажи пожалуйста в следующем сообщении как мне к тебе обращаться\n'), time_update(user_id=user_id)
+            await bot.send_message(message.from_user.id,'Здавствуй пользователь я бот Grek, как я вижу ты здесь впервые укажи пожалуйста в следующем сообщении как мне к тебе обращаться\n')
         else:
             boolean_variable = -1
             while boolean_variable == -1:
                 boolean_variable = [index for index in range(len(user_base)) if user_id_base[index] == user_id], await bot.send_message(message.from_user.id,f"Приветствую {user_name_base[boolean_variable]} чем я могу вам помочь ? \nДля получения справки ввидите команду /help"), time_update(user_id=user_id), await bot.send_message(message.from_user.id, "/start\n/help\n/dice\n/func_menu")
+            count_of_message(user_id)
         return user_id_base, user_name_base
-        count_of_message(user_id)
+
 
 def time_update(user_id):
     global user_base
@@ -81,7 +83,7 @@ def time_update(user_id):
             current_time = now.strftime("%d:%H:%M:%S")
             user_base.at[boolean_variable[0], 'last call'] = current_time
             user_base.at[boolean_variable[0], 'status'] = "Online"
-            user_base.to_excel("user_base.xlsx")
+            user_base.to_excel("User_base.xlsx")
     count_of_message(user_id)
 
 @dp.message_handler(commands=['help'])
@@ -99,13 +101,13 @@ async def process_help_command(message: types.Message):
 @dp.message_handler(commands=['func_menu'])
 async def func_menu(message: types.Message):
     user_id = message.from_user.id
+    admin_base = user_base['admin'].tolist()
     if ban_chek(user_id) != 1:
         user_id_base = user_base['id'].tolist()
         time_update(user_id=user_id)
         boolean_variable = -1
         while boolean_variable == -1:
             boolean_variable = [index for index in range(len(user_base)) if user_id_base[index] == user_id]
-        admin_base = user_base['admin'].tolist()
         if admin_base[boolean_variable[0]] == "NO":
             keyboard = types.InlineKeyboardMarkup()
             keyboard.add(types.InlineKeyboardButton(text="Какой сегодня день", callback_data="What day is today"))
@@ -154,7 +156,7 @@ async def Connect_administrator_rights(call: types.CallbackQuery):
 async def feature_list(call: types.CallbackQuery):
     user_id = call.from_user.id
     if ban_chek(user_id) != 1:
-        await call.message.answer("Вот что я умею:\nотвечать на ваше приветсттвие\nПоказывать погоду в указаном вами городе\nБросать кубик через команду /dice\nРассказывать какой сегодня день"), await call.message.answer("/start\n/help\n/dice\n/func_menu")
+        await call.message.answer("Вот что я умею:\nотвечать на ваше приветсттвие\nПоказывать погоду в указаном вами городе\nБросать кубик через команду /dice\nРассказывать какой сегодня день\nПодбирать Обои для вашего рабочего стола"), await call.message.answer("/start\n/help\n/dice\n/func_menu")
 
 
 @dp.callback_query_handler(text="Change nickname")
@@ -185,7 +187,8 @@ def multiple_replace(page, dictionary):
 async def processing_message(message: types.Message):
     global user_base, user_name_base, ch_log, con_adm_rig, Im_sear
     user_id = message.from_user.id
-    if ban_chek(user_id) != 1:
+    user_id_base = user_base['id'].tolist()
+    if ban_chek(user_id) != 1 or (user_id in user_id_base) == False:
         user_id_base = user_base['id'].tolist()
         user_name_base = user_base['name'].tolist()
         user_id = message.from_user.id
@@ -194,68 +197,88 @@ async def processing_message(message: types.Message):
             new_user = {'id': [user_id], 'name': [message.text], 'admin': ["NO"], 'message in 5 seconds': [0], 'ban': ["NO"]}
             df = pd.DataFrame(new_user, columns=['id', 'name', 'admin'])
             user_base = pd.concat([user_base, df], ignore_index=True, sort=False)
-            user_base.to_excel("user_base.xlsx")
+            user_base.to_excel("User_base.xlsx")
             await bot.send_message(message.from_user.id, "Отлично я успешно внёс вас в базу данных"), await bot.send_message(message.from_user.id, "/start\n/help\n/dice\n/func_menu"), time_update(user_id=user_id)
             count_of_message(user_id)
         elif ch_log == 1:
             user_base.at[index[0], 'name'] = message.text
-            user_base.to_excel("user_base.xlsx")
+            user_base.to_excel("User_base.xlsx")
             await bot.send_message(message.from_user.id, f"Отлично я успешно сменил ваш ник на {message.text}"), await bot.send_message(message.from_user.id, "/start\n/help\n/dice\n/func_menu")
             ch_log = 0
             count_of_message(user_id)
         elif con_adm_rig == 1:
             time_update(user_id=user_id)
             if message.text == ADMIN_PASS:
-                index = [index for index in range(len(user_base)) if user_id_base[index] == user_id]
-                user_base.at[index[0], 'admin'] = "YES", user_base.to_excel("user_base.xlsx")
+                user_base.at[index[0], 'admin'] = 'YES', user_base.to_excel("User_base.xlsx")
                 await bot.send_message(message.from_user.id,"Вы успешно получили права администратора"), await bot.send_message(message.from_user.id, "/start\n/help\n/dice\n/func_menu")
             else: await bot.send_message(message.from_user.id, "Неверный пароль")
             con_adm_rig = 0, time_update(user_id=user_id)
             count_of_message(user_id)
         elif Im_sear == 1:
+            user_base.at[index[0], 'ban'] = "YES"
             time_update(user_id=user_id)
-            images = 1
-            index = 1
+            images = 0
+            switch = 1
             num_page = 1
             keyword = message.text
             ip = await ip_database_manager()
             proxies = {'http': f'http://{ip}'}
-            while index == 1 and images != 11:
+            await bot.send_message(message.from_user.id, "Начинаю искать подходящие картинки")
+            while switch >= 1 and images != 10:
                 link = f'https://zastavok.net'
                 request = f'/search/{keyword}'
                 url = f'{link}/{request}/{num_page}/'
                 response = requests.get(url, proxies=proxies).text
                 if ('ничего не найдено' in response) == False:
-                    await bot.send_message(message.from_user.id,"Нашёл несколько интересных изображений по вашему запросу, сейчас перешлю вам")
                     page = BeautifulSoup(response, "lxml")
                     block = page.find('div', class_='main')
                     num = page.find('div', id='clsLink3')
-                    if images == 1:
-                        maximum_number_of_pages = num.find_all('a').__str__()
+                    maximum_number_of_pages = num.find_all('a').__str__()
+                    if maximum_number_of_pages[(maximum_number_of_pages.rfind("</a>,")) - 2] == '>':
                         maximum_number_of_pages = [int(maximum_number_of_pages[(maximum_number_of_pages.rfind("</a>,")) - 1]) if maximum_number_of_pages != '[]' else 1]
+                    else:
+                        maximum_number_of_pages = [int(maximum_number_of_pages[(maximum_number_of_pages.rfind("</a>,")) - 2:(maximum_number_of_pages.rfind("</a>,"))]) if maximum_number_of_pages != '[]' else 1]
                     Pictures = block.find_all('div', class_='short_full')
                     for image in Pictures:
                         chek = image.find('img').get('alt')
-                        if ((f'{keyword} ' in chek) or (f'{keyword.lower()} ' in chek) or (keyword == chek) or (keyword.lower() == chek)) and images != 11:
-                            image_link = image.find('a').get('href')
-                            download = requests.get(f'{link}{image_link}').text
-                            download_page = BeautifulSoup(download, 'lxml')
-                            download_blok = download_page.find('div', class_='image_data').find('div', class_='block_down')
-                            download_link = download_blok.find('a').get('href')
-                            image_bytes = requests.get(f'{link}{download_link}').content
-                            with open(f'picture/{chek}.jpg', 'wb') as file:
-                                file.write(image_bytes)
-                            with open(f'picture/{chek}.jpg', 'rb') as photo:
-                                await bot.send_photo(chat_id=message.chat.id, photo=photo)
-                            os.remove(f'picture/{chek}.jpg')
-                            images += 1
-                    if maximum_number_of_pages[0] == num_page:index = 0
-                    else:num_page += 1
+                        if ((f'{keyword} ' in chek) or (f'{keyword.lower()} ' in chek) or (keyword == chek) or (keyword.lower() == chek)):
+                            switch += 1
+                            break
+                    if switch == 1:
+                        switch = 0
+                    if switch > 0:
+                        if images == 0:
+                            await bot.send_message(message.from_user.id, "Нашёл несколько интересных изображений по вашему запросу, сейчас перешлю вам")
+                        for image in Pictures:
+                            if images == 10:
+                                break
+                            chek = image.find('img').get('alt')
+                            if ((f'{keyword} ' in chek) or (f'{keyword.lower()} ' in chek) or (keyword == chek) or (keyword.lower() == chek)) and images != 11:
+                                image_link = image.find('a').get('href')
+                                download = requests.get(f'{link}{image_link}').text
+                                download_page = BeautifulSoup(download, 'lxml')
+                                download_blok = download_page.find('div', class_='image_data').find('div', class_='block_down')
+                                download_link = download_blok.find('a').get('href')
+                                image_bytes = requests.get(f'{link}{download_link}').content
+                                with open(f'picture/{chek}.jpg', 'wb') as file:
+                                    file.write(image_bytes)
+                                with open(f'picture/{chek}.jpg', 'rb') as photo:
+                                    await bot.send_photo(chat_id=message.chat.id, photo=photo)
+                                os.remove(f'picture/{chek}.jpg')
+                                images += 1
+                        if maximum_number_of_pages[0] == num_page:index = 0
+                        else:num_page += 1
+                    else:
+                        await bot.send_message(message.from_user.id, "По вашему запросу ничего не найдено")
+                        switch = 0
                 else:
                     await bot.send_message(message.from_user.id, "По вашему запросу ничего не найдено")
-                    index = 0
+                    switch = 0
+            user_base.at[index[0], 'ban'] = "NO"
+            user_base.to_excel("User_base.xlsx")
             await bot.send_message(message.from_user.id, "/start\n/help\n/dice\n/func_menu")
             count_of_message(user_id)
+            Im_sear == 0
         elif message.text == "Привет" or message.text == "привет":
             status_base = user_base['status'].tolist()
             if status_base[index[0]] == 'Online':
@@ -302,16 +325,18 @@ def base_call(id, message):
 def ban_chek(id):
     global user_base
     user_id_base = user_base['id'].tolist()
-    ban_base = user_base['ban'].tolist()
-    boolean_variable = [index for index in range(len(user_base)) if user_id_base[index] == id]
-    if ban_base[boolean_variable[0]] != "YES":return 0
-    else:return 1
+    if (id in user_id_base) == True:
+        ban_base = user_base['ban'].tolist()
+        boolean_variable = [index for index in range(len(user_base)) if user_id_base[index] == id]
+        if ban_base[boolean_variable[0]] != "YES":return 0
+        else:return 1
+    else:return 0
 def count_of_message (id):
     global user_base
     user_id_base = user_base['id'].tolist()
     boolean_variable = [index for index in range(len(user_base)) if user_id_base[index] == id]
     user_base.at[boolean_variable[0], 'message in 5 seconds'] += 1
-    user_base.to_excel("user_base.xlsx")
+    user_base.to_excel("User_base.xlsx")
 
 
 async def spam_chek ():
@@ -326,10 +351,10 @@ async def spam_chek ():
                     user_base.at[index, 'ban'] = "YES"
                     await send_message(id_list[index], "Вы получили бан , бот с вами больше не общается")
                     user_base.at[index, 'message in 5 seconds'] = 0
-                    user_base.to_excel("user_base.xlsx")
+                    user_base.to_excel("User_base.xlsx")
                 else:
                     user_base.at[index, 'message in 5 seconds'] = 0
-                    user_base.to_excel("user_base.xlsx")
+                    user_base.to_excel("User_base.xlsx")
         await asyncio.sleep(5)
 
 
@@ -344,14 +369,22 @@ async def get_new_ip_list():
     ip_list_string = response[(response.find('UTC.\n') + 6):(response.find('</textarea>'))]
     for element in range(0, ip_list_string.count('\n')):
         ip_list.append(ip_list_string[0:ip_list_string.find('\n')])
+        ip_base.at[element, 'ip_list'] = ip_list[element]
+        print(ip_list[element])
         ip_list_string = ip_list_string[(ip_list_string.find('\n')) + 2:len(ip_list_string)]
+    ip_base.to_excel("ip_base.xlsx")
 
 
 async def ip_database_manager():
     global ip_list, element
-    if ip_list == [] or element == len(ip_list):
-        await get_new_ip_list()
     element += 1
+    ip_base.at[0,'number_ip'] =element
+    ip_base.to_excel("ip_base.xlsx")
+    if ip_list[0] == -1 or element == len(ip_list):
+        await get_new_ip_list()
+        element = -1
+        ip_base.at[0, 'number_ip'] = element
+        ip_base.to_excel("ip_base.xlsx")
     return ip_list[element]
 
 
@@ -370,18 +403,18 @@ async def status_chek():
             time_3 = int((time_base[index][0]) + (time_base[index][1]))
             if time_3 < current_time_date:
                 user_base.at[index, 'status'] = "offline"
-                user_base.to_excel("user_base.xlsx")
+                user_base.to_excel("User_base.xlsx")
             elif time_1 < current_time_minutes:
                 if (current_time_minutes - time_1 >= 30):
                     user_base.at[index, 'status'] = "offline"
-                    user_base.to_excel("user_base.xlsx")
+                    user_base.to_excel("User_base.xlsx")
                 elif time_2 < current_time_hour:
                     if (current_time_hour - time_2 >= 1):
                         user_base.at[index, 'status'] = "offline"
-                        user_base.to_excel("user_base.xlsx")
+                        user_base.to_excel("User_base.xlsx")
                 else:
                     user_base.at[index, 'status'] = "Online"
-                    user_base.to_excel("user_base.xlsx")
+                    user_base.to_excel("User_base.xlsx")
         await send_message(1258306656, "status_chek completed:" + str(now))
         now = datetime.now()
         print(f"status_chek completed: {str(now)}")
